requirements.txt
Unidecode~=1.4.0
pillow~=11.2.1